Adapt denunciation models and basic structure to new project
